Improve image export using our own image exporter
The custom exporter is based on html-to-image.
However, since we are exporting an SVG with only limited needs
for foreignObjects the code can be much simpler.

We can also use tricks from jointjs
to reduce the number of styles we override on the svg.
This makes the svg much more usable (still need to take care
of foreign objects before we can make it works).

We can also get the size of the actual graph more easily now.