Fix marker events
It's more correct to use mouse enter and leave rather than over and out.
The former should only trigger when the mouse enter and leave the whole element
(rather than it's decendents) and should not trigger when the mouse is moving in the element.

Somehow after updating the mouse over starts to trigger after the mouse out
causing the marker to be shown again immediately which appears as lingering to the user....