Tweak side panel focusing behavior
The old behavior is getting a bit annoy now that we can reliably focus on the editor
when the user clicks the editor webview. I'd still like to make it clear to the user
that the side panel is an integrate part of the editor,
which is achieved both by adding another button on our editor to activate the side panel
(that looks just like the one we put on on other editors)
and by automatically focus on the side panel
if we are bringing up the editor due to some action by the user (like pressing one of our buttons).

This way, the side panel should still activate when the user did something that will most likely need it, but would allow the user to do most normal editor things without having the side panel popping up all the time.